fix: clear the warning baseline to zero (TI-004)
64 gcc warnings and 60 clang warnings to none, on the set from BUILD.md
section 6, with -Werror now buildable (TYPEIT_WERROR=ON verified locally on
gcc; TI-020 turns it on in CI).

What was actually wrong, per review section 4:
- -Wreorder on the Input, Menu and Timer constructors: initialiser lists now
  follow declaration order. This is the class of bug where a member is
  initialised from another that has not been initialised yet.
- -Wparentheses in Text.cpp: the wrap condition is now explicitly grouped as
  the compiler was already grouping it.
- 48 signed/unsigned conversions, concentrated in InputLineEngine: fixed with
  explicit casts that preserve the existing conversion exactly, including
  where that conversion is itself dubious. `size() - 1` on an empty container
  still wraps to SIZE_MAX here; making the cast explicit documents it without
  changing it, and TI-033 replaces the engine outright.
- One missing field initialiser and two unused lambda captures.

Behaviour preservation is proved rather than asserted: a harness runs the old
and new Text::populate_text_lines over sixteen inputs -- the existing tests'
inputs plus the wrap boundary at 53, 54 and 55 characters, a lone space and a
lone newline -- and the line counts are identical.
TextTest.WrapConditionUnchangedAfterParenthesisation pins those numbers.

Two test files are touched, which the issue asks to avoid. The counters passed
to render_input_text were ints against a std::size_t parameter, and one
EXPECT_FLOAT_EQ compared an int-returning function against 0. The warning set
applies to the test target too, so zero warnings is unreachable otherwise. No
expected value changed and no assertion was weakened; the suite goes 61 to 62
tests, all passing.

docs/warning-baseline.txt is deleted, as TI-003 said it would be.

+++ include/data/GameState.h
@@ -18,8 +18,10 @@ struct GameState {
     static void toggle_info_display() { show_info = !show_info; }
 
     static void start_game_session() {
-        const std::string& path = GameOptions::text_radiobox_values_[GameOptions::selected_radiobox_text_];
-        const int time = GameOptions::time_radiobox_values_[GameOptions::selected_radiobox_time_];
+        const std::string& path =
+                GameOptions::text_radiobox_values_[static_cast<std::size_t>(GameOptions::selected_radiobox_text_)];
+        const int time =
+                GameOptions::time_radiobox_values_[static_cast<std::size_t>(GameOptions::selected_radiobox_time_)];
         if (FileTextSource::is_file_valid(path) && time > 0) {
             start_session = true;
         }
