+++ src/engines/InputLineEngine.cpp
@@ -13,13 +13,15 @@
 
 InputLineEngine::InputLineEngine(const Text& text_instance) :
     current_line_index_(0), text_instance_(text_instance),
-    total_input_lines_(text_instance_.get_text_lines_size() + 1, ftxui::hbox(ftxui::text(""))) {}
+    total_input_lines_(static_cast<std::size_t>(text_instance_.get_text_lines_size() + 1),
+                       ftxui::hbox(ftxui::text(""))) {}
 
 
 [[nodiscard]] int InputLineEngine::get_previous_lines_size() const {
-    return std::accumulate(
-            previous_input_lines_.begin(), previous_input_lines_.end(), 0,
-            [](const int sum, const ftxui::Elements& line) { return sum + std::distance(line.begin(), line.end()); });
+    return std::accumulate(previous_input_lines_.begin(), previous_input_lines_.end(), 0,
+                           [](const int sum, const ftxui::Elements& line) {
+                               return sum + static_cast<int>(std::distance(line.begin(), line.end()));
+                           });
 }
 
 
@@ -27,21 +29,24 @@ void InputLineEngine::go_to_new_line() {
     total_input_lines_[current_line_index_++] = ftxui::hbox(current_input_line_);
     previous_input_lines_.push_back(current_input_line_);
     current_input_line_.clear();
-    if (current_line_index_ < text_instance_.get_text_lines_size() - 1) {
+    if (current_line_index_ < static_cast<std::size_t>(text_instance_.get_text_lines_size() - 1)) {
         FocusPosition::y++;
     }
 }
 
 
 bool InputLineEngine::should_go_to_next_line(char next_character) const {
-    return next_character == ' ' && current_line_index_ < text_instance_.get_text_lines_size() - 1 &&
-           current_input_line_.size() >= text_instance_.get_text_line_size(current_line_index_);
+    return next_character == ' ' &&
+           current_line_index_ < static_cast<std::size_t>(text_instance_.get_text_lines_size() - 1) &&
+           current_input_line_.size() >=
+                   static_cast<std::size_t>(text_instance_.get_text_line_size(static_cast<int>(current_line_index_)));
 }
 
 
 bool InputLineEngine::should_finish_game() {
-    return current_line_index_ + 1 >= text_instance_.get_text_lines_size() &&
-           current_input_line_.size() + 1 >= text_instance_.get_text_line_size(current_line_index_);
+    return current_line_index_ + 1 >= static_cast<std::size_t>(text_instance_.get_text_lines_size()) &&
+           current_input_line_.size() + 1 >=
+                   static_cast<std::size_t>(text_instance_.get_text_line_size(static_cast<int>(current_line_index_)));
 }
 
 
@@ -62,14 +67,14 @@ void InputLineEngine::go_to_previous_line() {
     current_line_index_ ? current_line_index_-- : 0;
     current_input_line_ = previous_input_lines_[current_line_index_];
     previous_input_lines_.pop_back();
-    if (current_line_index_ < text_instance_.get_text_lines_size() - 2) {
+    if (current_line_index_ < static_cast<std::size_t>(text_instance_.get_text_lines_size() - 2)) {
         FocusPosition::y--;
     }
 }
 
 
 void InputLineEngine::remove_element(std::size_t input_text_size) {
-    if (input_text_size < get_previous_lines_size()) {
+    if (input_text_size < static_cast<std::size_t>(get_previous_lines_size())) {
         go_to_previous_line();
         return;
     }
@@ -78,13 +83,15 @@ void InputLineEngine::remove_element(std::size_t input_text_size) {
 
 
 bool InputLineEngine::should_add_element(std::size_t input_text_size) const {
-    const size_t previous_elements_size = get_previous_lines_size() + current_input_line_.size();
+    const size_t previous_elements_size =
+            static_cast<std::size_t>(get_previous_lines_size()) + current_input_line_.size();
     return input_text_size > previous_elements_size;
 }
 
 
 bool InputLineEngine::should_remove_element(std::size_t input_text_size) const {
-    const size_t previous_elements_size = get_previous_lines_size() + current_input_line_.size();
+    const size_t previous_elements_size =
+            static_cast<std::size_t>(get_previous_lines_size()) + current_input_line_.size();
     return input_text_size < previous_elements_size;
 }
 
@@ -103,8 +110,9 @@ void InputLineEngine::render_input_text(char next_character, std::size_t input_t
 // Transforms character to ftxui element.
 [[nodiscard]] ftxui::Element InputLineEngine::get_next_character(const char next_character) {
     ftxui::Element new_character = ftxui::text(std::string(1, next_character));
-    if (next_character == text_instance_.get_char_at_line_and_position(
-                                  current_line_index_, std::max(static_cast<int>(current_input_line_.size()), 0))) {
+    if (next_character ==
+        text_instance_.get_char_at_line_and_position(static_cast<int>(current_line_index_),
+                                                     std::max(static_cast<int>(current_input_line_.size()), 0))) {
         input_accuracy_.push_character_accuracy(true);
         return new_character | Style::input_text_color_good;
     }
